feat(backend): We actually store users now!
This commit introduces a *lot* in order to setup registration w.r.t. the
database. The major thing is how migrations work: We go through each
migration in the `migrations/scripts` folder, and run them against the
database. The ones we've executed are stored in our `migrations` table,
and we query against it to make sure we don't rerun those different
migrations again.

It's important to note that these migrations have to start with a number
so that they'll run in order.

It's a bit messy but it's also worth noting right now that our `Config`
class is actually also our data manager. We probably want to wrangle
this out.